requirements.txt
django-ckeditor==6.4.2
Pillow==9.1.1
django==4.0.4
django_countries